Titanic Survival Prediction
The Titanic dataset is a classic and widely-used dataset from Kaggle that contains data about passengers on the RMS Titanic, which sank in 1912. The main objective of this project is to analyze the data and build machine learning models to predict whether a passenger survived or not, based on features such as age, gender, fare, and travel class.